% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Random fish catch data}
\format{
A list of of 4 fish species
\itemize{
\item species species of fish caught
}
}
\source{
made up
}
\usage{
fish
}
\description{
Data of random types of fish (Description)
}
\keyword{datasets}
