% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_risk_score.R
\name{temp_risk_score}
\alias{temp_risk_score}
\title{Extreme Temperature Risk Score}
\usage{
temp_risk_score(
  temp,
  age = 200,
  temp_ideal = 20,
  highest_threshold = 38,
  medium_threshold = 32
)
}
\arguments{
\item{temp}{vector of daily temperatures (C)}

\item{age}{average tree age (years) default=200}

\item{temp_ideal}{ideal temperature for the tree species being considered (C) default=20}

\item{highest_threshold}{highest risk threshold temperature (C) default=38}

\item{medium_threshold}{medium risk threshold temperature (C) default=32}
}
\value{
risk, num_extremes
}
\description{
Our function takes a vector of daily temperatures and returns the risk
associated with the temperature vector and the number of extremes.
}
